feat: use click + ctrl to change reference element
